+++ fx/data.go
@@ -3,9 +3,9 @@ package fx
 type Record struct {
 	Date   string
 	Time   string
-	Open   float32
-	High   float32
-	Low    float32
-	Close  float32
+	Open   float64
+	High   float64
+	Low    float64
+	Close  float64
 	Volume uint64
 }
